Fix weird algorithm to use different utility function (#277)
* DDR fixed

* GEG fixed

* MRR fixed

* Fix no algorithm found in ALgorithmRunner

* Resolve issues around prepare method

* Move utility_function from algorithm options to configs

+++ api/ai/double_round_robin_algorithm/custom_models.py
@@ -2,7 +2,7 @@ from api.models.student import Student
 
 
 class Utility:
-    def __init__(self, value: int, student: Student, project_id: int):
+    def __init__(self, value: float, student: Student, project_id: int):
         self.value = value
         self.student = student
         self.project_id = project_id
